+++ CHANGELOG.rst
@@ -2,9 +2,25 @@
 Changelog
 =========
 
-Version 0.1
+Sources to write the changelog:
+- https://keepachangelog.com/en/1.0.0/
+- https://semver.org/
+
+Version 0.0
 ===========
 
-- Feature A added
-- FIX: nasty bug #1729 fixed
-- add your changes here!
+- Finishing the initial version of the project.
+- Missing:
+    - Tests
+    - Documentation
+    - Examples
+    - CI/CD
+    - Code quality checks
+    - Versioning
+    - Packaging
+    - Release notes
+- Missing Features:
+    - Experience replay.
+
+- Bugs:
+    - More cores/vectorized envs makes the training unstable.